fix: variation value input is now shown in variation page
